C: Fixed wrong length on udp data 32->16 bits, adding big endian conversion for network packet write

+++ eth_defs.h
@@ -24,6 +24,6 @@
 #define DEFAULT_SRC_MAC ((uint8_t[6]) { 0x00, 0x00, 0x01, 0, 0, 0})
 #define DEFAULT_DST_IP (uint32_t) 0x0
 #define DEFAULT_SRC_IP (uint32_t) 0x1
-#define DEFAULT_DST_PORT (uint32_t) 80
-#define DEFAULT_SRC_PORT (uint32_t) 80
+#define DEFAULT_DST_PORT (uint16_t) 80
+#define DEFAULT_SRC_PORT (uint16_t) 80
 #endif // ETH_DEFS_H
